Feature/rpn 1285 (#93)
* RPN-1285

* Reconciliation enabled for active blockchains

+++ integration/tezos/src/main/kotlin/com/rarible/protocol/union/integration/tezos/converter/TezosItemConverter.kt
@@ -72,7 +72,7 @@ object TezosItemConverter {
     ): RoyaltyDto {
         return RoyaltyDto(
             account = UnionAddressConverter.convert(source.account, blockchain),
-            value = source.value.toBigDecimal() // TODO TEZOS do we need some multiplier here?
+            value = source.value
         )
     }
 }
